resolved pr comments?
